+++ js/main.js
@@ -1,7 +1,9 @@
 // import { getGeneratingMarkupElements } from './helpers/generating-markup-elements.js';
 import { switchForm, submittingForm,  form, formValues, disabledItem, disabledAdress } from './helpers/form.js';
 import { getSelectedItem, getSelectedPrice, getSelectedTime } from './helpers/get-selected-item.js';
-import { getMapDisplay } from './helpers/map.js';
+// import { getMapDisplay } from './helpers/map.js';
+// import { getMap } from './helpers/getMap.js';
+import { getMap} from './map/get-map.js';
 
 // getGeneratingMarkupElements(); // генерация фейковых данных
 switchForm(); // переключение формы в (не) активное состояние
@@ -14,4 +16,5 @@ form.addEventListener('submit', (evt) => submittingForm(evt)); //событие 
 getSelectedPrice(formValues); // синхронизация полей тип жилья и цена за ночь
 getSelectedTime(formValues.checkInTime , formValues.checkOutTime); // синхронизация полей день заезда
 getSelectedTime(formValues.checkOutTime, formValues.checkInTime); //и день выезда
-getMapDisplay(); //отображения карты
+// getMapDisplay(); //отображения карты
+getMap();
